errors fixed

+++ include/configuration.hpp
@@ -25,5 +25,5 @@ struct configuration
     hash_type hash{hash_type::kmer};
     uint8_t window_size{20u};
     uint8_t s{6u};
-    uint t{3u};
+    uint8_t t{3u};
 };
